Created a custom JS DATE API hook
